refator: 메인 페이지 리팩토링
