+++ app/createTrip/createTrip.tsx
@@ -1,5 +1,5 @@
 import { ScrollView } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
+import { SafeAreaView } from "react-native";
 
 import styles from "./styles";
 import Content from "@/components/createTrip/content/content";
