+++ components/createSchedule/calendar/calendar.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from "react";
 import { View, Text, TouchableOpacity, Modal } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
+import { SafeAreaView } from "react-native";
 import moment from "moment";
 import "moment/locale/ko";
 import styles from "./styles";
