+++ mock/data/common/main.ts
@@ -1,0 +1,50 @@
+import sokcho from "@/assets/images/sokcho.png";
+import chuncheon from "@/assets/images/chuncheon.png";
+import daegu from "@/assets/images/daegu.png";
+import jeans from "@/assets/images/jeans.png";
+
+export const historyData = [
+  {
+    id: "1",
+    destination: sokcho,
+    mainDescription: "속초 여행",
+    subDescription: "2024-06-18~2024-06-19",
+    hashTag: "#속초 해산물 #속초 포켓몬빵 #속초 헤엄",
+  },
+  {
+    id: "2",
+    destination: chuncheon,
+    mainDescription: "춘천 여행",
+    subDescription: "2024-08-14~2024-06-19",
+    hashTag: "#200일 #애인 #달갈비 #자전거",
+  },
+  {
+    id: "3",
+    destination: chuncheon,
+    mainDescription: "춘천 여행",
+    subDescription: "2024-08-14~2024-06-19",
+    hashTag: "#200일 #애인 #달갈비 #자전거",
+  },
+  {
+    id: "3",
+    destination: chuncheon,
+    mainDescription: "춘천 여행",
+    subDescription: "2024-08-14~2024-06-19",
+    hashTag: "#200일 #애인 #달갈비 #자전거",
+  },
+  {
+    id: "3",
+    destination: chuncheon,
+    mainDescription: "춘천 여행",
+    subDescription: "2024-08-14~2024-06-19",
+    hashTag: "#200일 #애인 #달갈비 #자전거",
+  },
+];
+
+export const upCommingData = {
+  destination: "대구",
+  thumbnail: daegu,
+  day: "D-19",
+  nickname: "청바지",
+  groupThumbnail: jeans,
+};
