+++ mock/data/createSchedule.ts/createSchedule.ts
@@ -1,0 +1,31 @@
+import shin from "@/assets/images/shin.png";
+import chul from "@/assets/images/chul.png";
+import mang from "@/assets/images/mang.png";
+import hoon from "@/assets/images/hoon.png";
+
+export const inviteData = [
+  {
+    id: 1,
+    name: "신짱구",
+    lead: true,
+    image: shin,
+  },
+  {
+    id: 2,
+    name: "김철수",
+    lead: false,
+    image: chul,
+  },
+  {
+    id: 3,
+    name: "맹구",
+    lead: false,
+    image: mang,
+  },
+  {
+    id: 4,
+    name: "훈이",
+    lead: false,
+    image: hoon,
+  },
+];
