+++ types/createSchedule/createSchedule.ts
@@ -1,0 +1,9 @@
+interface inviteData {
+  id: number;
+  name: string;
+  lead: boolean;
+  image: any;
+}
+export interface inviteDataState {
+  data: inviteData[];
+}
